refactor: Comment 패키지 수정
