+++ src/main/java/com/travel/role/domain/comment/service/CommentService.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.travel.role.domain.comment.repository.CommentRepository;
-import com.travel.role.domain.comment.dto.CommentListResDTO;
-import com.travel.role.domain.comment.dto.CommentReqDTO;
-import com.travel.role.domain.comment.dto.CommentResDTO;
+import com.travel.role.domain.comment.dto.response.CommentListResDTO;
+import com.travel.role.domain.comment.dto.request.CommentReqDTO;
+import com.travel.role.domain.comment.dto.response.CommentResDTO;
 import com.travel.role.domain.comment.entity.Comment;
 import com.travel.role.global.exception.comment.CommentInfoNotFoundException;
 import com.travel.role.domain.room.repository.RoomParticipantRepository;
